+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Version - 0]
+## fix [v0.50.0] 2024-3-1
+- admin report not show when detect role
+
 
 ## improve [v0.49.1] 2024-2-16